22w06a

+++ net/minecraft/entity/mob/ZombieEntity.java
@@ -171,7 +171,7 @@ extends HostileEntity {
     @Override
     protected int getXpToDrop(PlayerEntity player) {
         if (this.isBaby()) {
-            this.experiencePoints = (int)((float)this.experiencePoints * 2.5f);
+            this.experiencePoints = (int)((double)this.experiencePoints * 2.5);
         }
         return super.getXpToDrop(player);
     }
